requirements.txt
flask
python-pptx
python-docx
scikit-learn
joblib
gunicorn